fix(switchdash): break the import cycle that leaves view-registry half-built
view-registry builds `views` from every view module, and two modules the view
modules themselves reach imported the registry's value back: navigation-provider
(via useWorkspaceSlots) and the command registry (reachable through the modal
tree). Either edge lets the registry evaluate while a view binding it reads is
still in its temporal dead zone, so the renderer throws "Cannot access
'remoteHostsView' before initialization" — deterministically in CI, and not at
all locally, since which module wins the race is load-order dependent.

Move useWorkspaceSlots to its own module, pass the view map into
setupViewCommandProvider from the entry point rather than importing it, and give
callers that only need the active view id a registry-free hook so the sessions
view stops reaching the registry through use-is-active-session.

No runtime path now leads back into view-registry; only the entry point and
workspace-slots import its value, and neither is reachable from a view.
