+++ dash/apps/switchdash-desktop/src/renderer/app/app-menu-events.tsx
@@ -1,7 +1,8 @@
 import { useEffect } from 'react';
 import { toast } from '@renderer/lib/hooks/use-toast';
 import { events, rpc } from '@renderer/lib/ipc';
-import { useNavigate, useWorkspaceSlots } from '@renderer/lib/layout/navigation-provider';
+import { useNavigate } from '@renderer/lib/layout/navigation-provider';
+import { useWorkspaceSlots } from '@renderer/lib/layout/workspace-slots';
 import { toggleSettingsView } from '@renderer/lib/layout/settings-toggle';
 import { useShowModal } from '@renderer/lib/modal/modal-provider';
 import {
