+++ dash/apps/switchdash-desktop/src/renderer/features/sessions/hooks/use-is-active-session.ts
@@ -1,7 +1,10 @@
-import { useParams, useWorkspaceSlots } from '@renderer/lib/layout/navigation-provider';
+import { useCurrentViewId, useParams } from '@renderer/lib/layout/navigation-provider';
 
 export function useIsActiveSession(sessionId: string): boolean {
-  const { currentView } = useWorkspaceSlots();
+  // Deliberately not useWorkspaceSlots: this hook is reachable from the session
+  // view, so loading the view registry here would make the registry depend on
+  // its own members being initialised first.
+  const currentView = useCurrentViewId();
   const { params } = useParams('session');
   return currentView === 'session' && params.sessionId === sessionId;
 }
