+++ dash/apps/switchdash-desktop/src/renderer/features/sidebar/left-sidebar.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import {
   isCurrentView,
   useNavigate,
-  useWorkspaceSlots,
 } from '@renderer/lib/layout/navigation-provider';
+import { useWorkspaceSlots } from '@renderer/lib/layout/workspace-slots';
 import { BoundShortcut } from '@renderer/lib/ui/shortcut';
 import { cn } from '@renderer/utils/utils';
 import { ServersSidebarSection } from '../switch-servers/ServersSidebarSection';
