+++ dash/apps/switchdash-desktop/src/renderer/features/sidebar/session-item.tsx
@@ -10,8 +10,8 @@ import { SessionSidebarTrailingSlot } from '@renderer/features/sidebar/session-s
 import {
   useNavigate,
   useParams,
-  useWorkspaceSlots,
 } from '@renderer/lib/layout/navigation-provider';
+import { useWorkspaceSlots } from '@renderer/lib/layout/workspace-slots';
 import { useShowModal } from '@renderer/lib/modal/modal-provider';
 import { sidebarStore } from '@renderer/lib/stores/app-state';
 import { cn } from '@renderer/utils/utils';
