+++ dash/apps/switchdash-desktop/src/renderer/main.tsx
@@ -1,11 +1,11 @@
 import ReactDOM from 'react-dom/client';
-import { setupNavigationGuards } from '@renderer/app/view-registry';
+import { setupNavigationGuards, views } from '@renderer/app/view-registry';
 import { prefetchAppSettingsKey } from '@renderer/features/settings/use-app-settings-key';
 import './index.css';
 import 'devicon/devicon.min.css';
 import 'katex/dist/katex.min.css';
 import { setupAppCommandProvider } from '@renderer/lib/commands/app-commands';
-import { setupViewCommandProvider } from '@renderer/lib/commands/registry';
+import { setupViewCommandProvider, type ViewCommandProviders } from '@renderer/lib/commands/registry';
 import { wireExternalLinkRequests } from '@renderer/lib/external-link-requests';
 import { rpc } from '@renderer/lib/ipc';
 import { viewStateCache } from '@renderer/lib/stores/view-state-cache';
@@ -35,7 +35,7 @@ async function bootstrap() {
   setupNavigationGuards();
   if (navResult) appState.navigation.restoreSnapshot(navResult);
   setupAppCommandProvider();
-  setupViewCommandProvider();
+  setupViewCommandProvider(views as unknown as ViewCommandProviders);
   if (sidebarResult) {
     appState.sidebar.restoreSnapshot(sidebarResult as Partial<SidebarSnapshot>);
   } else {
